Add testing for the utils surrounding dynamo_timed
Summary: This will make it easier to verify that we don't break these utilities for the refactor in https://github.com/pytorch/pytorch/pull/139849.
It's one giant test. I can split it into multiple for better readability if ppl prefer that. My rationale for the giant test is that I found I was just resetting compilation and recompiling the same thing many times, which was slow and wasteful.

Test Plan: The new tests

ghstack-source-id: f56f705379d5cedc80de247d5b860b4259604dce
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140094